Feat
Imported Search component. Added fullList into state in order to create a combined list of top headline stories and articles for the search page to be able to parse through.
Renamed spareUrl to better reflect use of topStoriesUrl used in addtional fetch.
Removed quite a bit of commented out code.

Refactored goToArticle function to deal with the complication of needing to work with the HeadlineStories component, Search Component, Cards component. Used a switch to decide what to do by the type sent. This allows each of these components to render an article on click.

Created setArticlesForSearch function to populate the fullList array with the combination of both the headlines array and articles array.

Created Route for SearchPage component and passed in necessary props.